+++ bika/lims/browser/analysisrequest/__init__.py
@@ -25,7 +25,6 @@ from .add2 import AnalysisRequestAddView  # noqa: F401
 from .add2 import AnalysisRequestManageView  # noqa: F401
 from .add2 import ajaxAnalysisRequestAddView  # noqa: F401
 from .analysisrequests import AnalysisRequestsView
-from .analysisrequests import QueuedAnalysisRequestsCount
 from .invoice import InvoicePrintView
 from .invoice import InvoiceView
 from .manage_analyses import AnalysisRequestAnalysesView