+++ bika/lims/upgrade/v01_03_000.py
@@ -14,14 +14,33 @@ from bika.lims.catalog.analysisrequest_catalog import \
 from bika.lims.config import PROJECTNAME as product
 from bika.lims.upgrade import upgradestep
 from bika.lims.upgrade.utils import UpgradeUtils
+from plone.portlets.interfaces import IPortletType
+from zope import component
 
 version = '1.3.0'  # Remember version number in metadata.xml and setup.py
 profile = 'profile-{0}:default'.format(product)
+
+PORTLETS_TO_PURGE = [
+    'accreditation-pt',
+    'login',
+    'news',
+    'events',
+    'Calendar',
+    'portlet_department_filter-pt',
+    'portlet_to_be_sampled-pt',
+    'portlet_to_be_preserved-pt',
+    'portlet_late_analyses-pt',
+    'portlet_pending_orders-pt',
+    'portlet_sample_due-pt',
+    'portlet_to_be_verified-pt',
+    'portlet_verified-pt'
+]
 
 
 @upgradestep(product, version)
 def upgrade(tool):
     portal = tool.aq_inner.aq_parent
+    setup = portal.portal_setup
     ut = UpgradeUtils(portal)
     ver_from = ut.getInstalledVersion(product)
 
@@ -34,11 +53,59 @@ def upgrade(tool):
 
     # -------- ADD YOUR STUFF BELOW --------
 
+    # Remove QC reports and gpw dependency
+    # https://github.com/senaite/senaite.core/pull/1058
+    remove_qc_reports(portal)
+
+    # Remove updates notification viewlet
+    # https://github.com/senaite/senaite.core/pull/1059
+    setup.runImportStepFromProfile(profile, 'viewlets')
+
+    # Remove old portlets except the navigation portlet
+    # https://github.com/senaite/senaite.core/pull/1060
+    purge_portlets(portal)
+
     # Setup the partitioning system
     setup_partitioning(portal)
 
     logger.info("{0} upgraded to version {1}".format(product, version))
     return True
+
+
+def remove_qc_reports(portal):
+    """Removes the action Quality Control from Reports
+    """
+    logger.info("Removing Reports > Quality Control ...")
+    ti = portal.reports.getTypeInfo()
+    actions = map(lambda action: action.id, ti._actions)
+    for index, action in enumerate(actions, start=0):
+        if action == 'qualitycontrol':
+            ti.deleteActions([index])
+            break
+    logger.info("Removing Reports > Quality Control [DONE]")
+
+
+def purge_portlets(portal):
+    """Remove old portlets. Leave the Navigation portlet only
+    """
+    logger.info("Purging portlets ...")
+
+    def remove_portlets(context_portlet):
+        mapping = portal.restrictedTraverse(context_portlet)
+        for key in mapping.keys():
+            if key not in PORTLETS_TO_PURGE:
+                logger.info("Skipping portlet: '{}'".format(key))
+                continue
+            logger.info("Removing portlet: '{}'".format(key))
+            del mapping[key]
+
+    remove_portlets("++contextportlets++plone.leftcolumn")
+    remove_portlets("++contextportlets++plone.rightcolumn")
+
+    # Reimport the portlets profile
+    setup = portal.portal_setup
+    setup.runImportStepFromProfile(profile, 'portlets')
+    logger.info("Purging portlets [DONE]")
 
 
 def setup_partitioning(portal):
@@ -120,3 +187,4 @@ def add_index(portal, catalog_id, index_name, index_attribute, index_metatype):
     logger.info("Indexing new index '{}' ...".format(index_name))
     catalog.manage_reindexIndex(index_name)
     logger.info("Indexing new index '{}' [DONE]".format(index_name))
+